Tesztesetek ms-os eltérési fail hibájának javítása, Car-fleet-category-equipment-status integrációs teszt elkészítve.

+++ webapp/backend/database/migrations/2024_11_17_180100_create_categories_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->unsignedTinyInteger('category_class');
-            $table->unsignedTinyInteger('motor_power');
+            $table->unsignedInteger('motor_power');
         });
     }
     public function down(): void
